terraform: when promoting non-CBD to CBD, mark the config as such
Fixes #10439

When a CBD resource depends on a non-CBD resource, the non-CBD resource
is auto-promoted to CBD. This was done in
cf3a259cd9fd78bb9f2bc0949866b0708a48c10e. This PR makes it so that we
also set the config CBD to true. This causes the proper runtime
execution behavior to occur where we depose state and so on.

So in addition to simple graph edge tricks we also treat the non-CBD
resources as CBD resources.

As part of this process, I realized that the new graph doesn't support
this at all since there was no context test covering it! I added a
context test. This test also reproduced the error message in #10439
withou the config setting logic, so it shows it fixes that as well.